requirements.txt
requests>=2.25.0
Flask>=2.0.1
openpyxl>=3.0.7
python-dotenv>=0.19.0 